Implement integration tests for HlsSamplerPanel
Additionally, change ints to longs to avoid overflows and update sonar url

+++ src/main/java/com/blazemeter/jmeter/hls/logic/HlsSampler.java
@@ -101,12 +101,12 @@ public class HlsSampler extends HTTPSampler {
     String bandwidth = getPropertyAsString(CUSTOM_BANDWIDTH_PROPERTY_NAME);
     return BandwidthSelector
         .fromStringAndCustomBandwidth(getPropertyAsString(BANDWIDTH_TYPE_PROPERTY_NAME),
-            bandwidth != null && !bandwidth.isEmpty() ? Integer.valueOf(bandwidth) : null);
+            bandwidth != null && !bandwidth.isEmpty() ? Long.valueOf(bandwidth) : null);
   }
 
   public void setBandwidthSelector(BandwidthSelector selector) {
     setProperty(BANDWIDTH_TYPE_PROPERTY_NAME, selector.getName());
-    Integer bandwidth = selector.getCustomBandwidth();
+    Long bandwidth = selector.getCustomBandwidth();
     setProperty(CUSTOM_BANDWIDTH_PROPERTY_NAME, bandwidth != null ? bandwidth.toString() : null);
   }
